requirements.txt
pytest-playwright
progress
pytest
black
fire